secretr_app1/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
psycopg2-binary==2.9.10
requests==2.32.4
